+++ src/modules/bottom-sheet/bottom-sheet.types.ts
@@ -36,4 +36,7 @@ export interface IProps {
 
   /** If true, renders the header as fixed. */
   stickyHeader?: boolean;
+
+  /** If true, pressing the back button will close the bottom sheet. */
+  shouldCloseOnBackButtonPress?: boolean;
 }