+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.0.52
 
 - Bottom Sheet - Added prop `stickyHeader` to toggle between sticky and static header
+- Bottom Sheet - Added prop `shouldCloseOnBackButtonPress` to close the bottom sheet on back button press
 
 ---